Add functionality to read encrypted data from file (#3)
This change adds the functionality that the encrypted data from etcd can be read from a file.

In practice we have seen that the tool can fail with the following error message when a very large object (for example a large `Secret`) is pasted as the input:

```
Failed to decode etcd value: illegal base64 data at input byte 4093
```

It turns out that this is due to the terminal input buffer size limit, which is 4096 bytes on
GNU/Linux (look for N_TTY_BUF_SIZE in the kernel sources): https://groups.google.com/g/golang-nuts/c/ndh-1wdsWYs/m/Watbhx8JAwAJ

When such a case is detected, then a message is printed and the user should put his values in a file called `secretvalue`. The program will then read from that file if it is present and skip asking for the encrypted data.

The change also updates the README with the current location of the EncryptionConfig in newer OCP versions.